test(administration): split extension tooling report spec below the line limit
report.spec.ts crossed the 500-line ADR threshold; the check-report,
guidance, and setup-report suites move into report.spec/ with shared
fixture factories in helpers.ts, per
adr/2026-05-06-split-large-administration-test-files.md.
